added time to events, minor fixes

+++ users_app/models.py
@@ -11,7 +11,7 @@ def generateQrData():
 
 class Team(models.Model):
     name = models.CharField(max_length=100)
-    balance = models.IntegerField(default=0)
+    balance = models.FloatField(default=0)
     tutor = models.OneToOneField("User", related_name="tutor", null=True, blank=True, on_delete=models.DO_NOTHING)
 
     def __str__(self):
@@ -37,3 +37,4 @@ class User(AbstractUser):
     @staticmethod
     def all_users():
         return User.objects.all()
+
